+++ rpc-state-reader/src/utils.rs
@@ -1,21 +1,17 @@
 use std::{
     collections::HashMap,
-    error::Error,
-    fs::{self, File},
+    fs::{self},
     io::{self, Read},
-    path::{Path, PathBuf},
+    path::PathBuf,
     sync::{OnceLock, RwLock},
     time::Instant,
 };
 
-use blockifier::{
-    execution::{call_info::CallInfo, contract_class::CompiledClassV1},
-    transaction::objects::TransactionExecutionInfo,
-};
+use blockifier::execution::contract_class::CompiledClassV1;
 use cairo_lang_starknet_classes::contract_class::{ContractClass, ContractEntryPoints};
 use cairo_lang_utils::bigint::BigUintAsHex;
 use cairo_native::{executor::AotContractExecutor, OptLevel};
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use starknet::core::types::{LegacyContractEntryPoint, LegacyEntryPointsByType};
 use starknet_api::{
     contract_class::EntryPointType,
@@ -30,12 +26,6 @@ pub struct MiddleSierraContractClass {
     pub sierra_program: Vec<BigUintAsHex>,
     pub contract_class_version: String,
     pub entry_points_by_type: ContractEntryPoints,
-}
-
-#[derive(Debug, Serialize)]
-struct EntryPointExecution {
-    class_hash: ClassHash,
-    selector: EntryPointSelector,
 }
 
 static AOT_PROGRAM_CACHE: OnceLock<RwLock<HashMap<ClassHash, AotContractExecutor>>> =
@@ -165,64 +155,4 @@ pub fn get_casm_compiled_class(class: ContractClass, _class_hash: ClassHash) -> 
 
 pub fn bytecode_size(data: &[BigUintAsHex]) -> usize {
     data.iter().map(|n| n.value.to_bytes_be().len()).sum()
-}
-
-pub fn save_entry_point_execution(
-    path: &Path,
-    executions: Vec<(u64, TransactionExecutionInfo)>,
-) -> Result<(), Box<dyn Error>> {
-    let mut block_executions: HashMap<u64, Vec<HashMap<String, _>>> = HashMap::new();
-
-    for (block_number, execution) in executions {
-        let mut tx_execution: HashMap<String, _> = HashMap::new();
-
-        if let Some(call) = execution.validate_call_info {
-            tx_execution.insert(
-                "validate_call_info".to_string(),
-                get_inner_class_executions(call),
-            );
-        }
-        if let Some(call) = execution.execute_call_info {
-            tx_execution.insert(
-                "execute_call_info".to_string(),
-                get_inner_class_executions(call),
-            );
-        }
-        if let Some(call) = execution.fee_transfer_call_info {
-            tx_execution.insert(
-                "fee_transfer_call_info".to_string(),
-                get_inner_class_executions(call),
-            );
-        }
-
-        block_executions
-            .entry(block_number)
-            .or_insert_with(Vec::new)
-            .push(tx_execution);
-    }
-
-    let file = File::create(path)?;
-    serde_json::to_writer_pretty(file, &block_executions)?;
-
-    Ok(())
-}
-
-fn get_inner_class_executions(call: CallInfo) -> Vec<EntryPointExecution> {
-    // class hash can initially be None, but it is always added before execution
-    let class_hash = call.call.class_hash.unwrap();
-
-    let mut classes = call
-        .inner_calls
-        .into_iter()
-        .flat_map(get_inner_class_executions)
-        .collect::<Vec<_>>();
-
-    let top_class = EntryPointExecution {
-        class_hash,
-        selector: call.call.entry_point_selector,
-    };
-
-    classes.push(top_class);
-
-    classes
 }
